pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "django-fox"
version = "0.1.0"
authors = ["Thomas Baquet <thomas bkfox net>"]
description = "Django tools."
readme = "README.md"
packages = [
    { include = "fox/caps" },
    { include = "fox/etl" },
    { include = "fox/utils" }
]

[tool.poetry.dependencies]
Django = "^4.1"
django-model-utils = "^4.3"
django-filter = "^22.1"
djangorestframework = "^3.14"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1"
black = "^23.3"
ruff = "^0.0"
pytest = "^7.3"
pytest-django = "^4.5"

[tool.poetry.group.etl]
optional = true

[tool.poetry.group.etl.dependencies]
pandas = "^2.0"
